move the socket.io related logic into a separate file
