+++ api/src/index.js
@@ -1,4 +1,5 @@
 const app = require('./server')
+require('./socket')
 
 const port = process.env.PORT || 3000
